Revert "removed some warnings"
This reverts commit 55be88c76483792539a3d682a328c41b5f6ce0a6.

+++ tests/DAG.cpp
@@ -348,8 +348,8 @@ BOOST_AUTO_TEST_CASE(DAG1) {
         BOOST_CHECK( intersec == supposed_intersec );
     }
 
-    for (int  i = 0; i < static_cast<int>(graph.n); i++) {
-        for (int j = 0; j < static_cast<int>(graph.n); j++) {
+    for (unsigned i = 0; i < graph.n; i++) {
+        for (unsigned j = 0; j< graph.n; j++) {
             std::unordered_set<int> desc = graph.descendants(i);
             std::unordered_set<int> anc = graph.ancestors(j);
             std::unordered_set<int> intersec = get_intersection(desc,anc);
